# src/lumio-web/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Lumio API",
    "description": "Offline digitale nalatenschap beheer — Lumio",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/digitaal-bezit/accounts": {
      "get": {
        "tags": [
          "DigitaalBezit"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DigitaalAccountResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DigitaalAccountResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DigitaalAccountResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DigitaalBezit"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigitaalAccountUpsertRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DigitaalAccountUpsertRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DigitaalAccountUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DigitaalAccountResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigitaalAccountResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigitaalAccountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/accounts/{id}": {
      "put": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigitaalAccountUpsertRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DigitaalAccountUpsertRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DigitaalAccountUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DigitaalAccountResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigitaalAccountResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigitaalAccountResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/wachtwoorden": {
      "get": {
        "tags": [
          "DigitaalBezit"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WachtwoordEntryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WachtwoordEntryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WachtwoordEntryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DigitaalBezit"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WachtwoordEntryCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WachtwoordEntryCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WachtwoordEntryCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordEntryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordEntryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/wachtwoorden/{id}/ontsluitel": {
      "get": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordOntsluitelResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordOntsluitelResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordOntsluitelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/wachtwoorden/{id}": {
      "put": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WachtwoordEntryUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WachtwoordEntryUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WachtwoordEntryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordEntryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordEntryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WachtwoordEntryResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/wachtwoorden/importeren": {
      "post": {
        "tags": [
          "DigitaalBezit"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Bestand"
                ],
                "type": "object",
                "properties": {
                  "Bestand": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "Bestand": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/crypto": {
      "get": {
        "tags": [
          "DigitaalBezit"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CryptoWalletResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CryptoWalletResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CryptoWalletResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DigitaalBezit"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoWalletUpsertRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoWalletUpsertRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoWalletUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/crypto/{id}": {
      "put": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoWalletUpsertRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoWalletUpsertRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoWalletUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/digitaal-bezit/crypto/{id}/ontsluitel": {
      "get": {
        "tags": [
          "DigitaalBezit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletOntsluitelResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletOntsluitelResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CryptoWalletOntsluitelResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CryptoWalletResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "walletNaam": {
            "type": "string",
            "nullable": true
          },
          "cryptoType": {
            "type": "string",
            "nullable": true
          },
          "walletAdres": {
            "type": "string",
            "nullable": true
          },
          "exchange": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DigitaalAccountResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "platformNaam": {
            "type": "string",
            "nullable": true
          },
          "categorie": {
            "type": "string",
            "nullable": true
          },
          "gebruikersnaam": {
            "type": "string",
            "nullable": true
          },
          "emailAdres": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "gewensteActie": {
            "type": "string",
            "nullable": true
          },
          "overdrachtAan": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DigitaalAccountUpsertRequest": {
        "type": "object",
        "properties": {
          "platformNaam": {
            "type": "string",
            "nullable": true
          },
          "categorie": {
            "type": "string",
            "nullable": true
          },
          "gebruikersnaam": {
            "type": "string",
            "nullable": true
          },
          "emailAdres": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "gewensteActie": {
            "type": "string",
            "nullable": true
          },
          "overdrachtAan": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WachtwoordEntryCreateRequest": {
        "type": "object",
        "properties": {
          "naam": {
            "type": "string",
            "nullable": true
          },
          "gebruikersnaam": {
            "type": "string",
            "nullable": true
          },
          "wachtwoord": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WachtwoordEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "naam": {
            "type": "string",
            "nullable": true
          },
          "gebruikersnaam": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          },
          "hasPassword": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "WachtwoordEntryUpdateRequest": {
        "type": "object",
        "properties": {
          "naam": {
            "type": "string",
            "nullable": true
          },
          "gebruikersnaam": {
            "type": "string",
            "nullable": true
          },
          "nieuwWachtwoord": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WachtwoordOntsluitelResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "naam": {
            "type": "string",
            "nullable": true
          },
          "gebruikersnaam": {
            "type": "string",
            "nullable": true
          },
          "wachtwoord": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "notities": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}